refactor: simplify architecture
basic idea of architecture is to divide program into server (requests handler), service (business logic), storage (DB) and core (basic structs, interfaces)
