+++ config.yaml
@@ -12,6 +12,7 @@ concurrency_limit:
 
 rate_limit:
   5
+
 token_max_time:
   60
